improved tests by removing duplicate and moving test to correct folder
