+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/container-wrapper-component-test.ts
@@ -4,6 +4,8 @@ import type { StixContainer } from '../../../../src/types/stix-2-1-sdo';
 import { testBundleObject, testExecutor } from './playbook-components-test-utils';
 import type { StixRelation } from '../../../../src/types/stix-2-1-sro';
 import { playbookBundleElementsToApply } from '../../../../src/modules/playbook/playbook-types';
+import type { StixBundle } from '../../../../src/types/stix-2-1-common';
+import { STIX_EXT_OCTI } from '../../../../src/types/stix-2-1-extensions';
 
 const componentConfig = (config: Partial<ContainerWrapperConfiguration>) => {
   return {
@@ -79,5 +81,122 @@ describe('PLAYBOOK_CONTAINER_WRAPPER_COMPONENT', () => {
     const reportResult = result.bundle.objects
       .find((element) => element.type === 'report') as StixContainer;
     expect(reportResult.object_refs).toEqual([CAMPAIGN_ID, REL_ID]);
+  });
+
+  it('should wrap Incident in Incident Response container', async () => {
+    const INCIDENT_ID = 'incident--c6c2b96d-fe70-5099-a033-87cbfe2d6be2';
+
+    const incidentBundleObject = testBundleObject({
+      id: INCIDENT_ID,
+      type: 'Incident',
+      extension: {
+        created_at: '2025-02-25T08:13:45.863Z',
+        updated_at: '2025-05-09T10:03:11.288Z',
+        is_inferred: false,
+        granted_refs: ['34a50091-acd7-5b12-88f4-086155cf40d4'],
+        creator_ids: ['88ec0c6a-13ce-5e39-b486-354fe4a7084f'],
+      } as unknown as StixRelation });
+
+    const container_wrapper_component_bundle: StixBundle = {
+      id: '1c775f39-6cea-4b14-92f8-7843d2443af7',
+      spec_version: '2.1',
+      type: 'bundle',
+      objects: [
+        {
+          ...incidentBundleObject,
+          external_references: [
+            {
+              source_name: 'upload_file',
+              external_id: 'upload_file_example.pdf',
+            },
+          ],
+          severity: 'high',
+          revoked: false,
+          confidence: 100,
+          lang: 'en',
+          name: 'Test Incident',
+          description: '',
+
+        },
+      ],
+    } as unknown as StixBundle;
+    const dataInstanceId = INCIDENT_ID;
+    const playbookNode = {
+      id: 'fb1d0a03-eb72-426a-ab48-505a2ca399d0',
+      name: 'Container wrapper',
+      position: {
+        x: 0,
+        y: 150,
+      },
+      component_id: 'PLAYBOOK_CONTAINER_WRAPPER_COMPONENT',
+      configuration: {
+        container_type: 'Case-Incident',
+        applyToElements: playbookBundleElementsToApply.onlyMain.value,
+        copyFiles: false,
+        newContainer: false,
+        caseTemplates: [],
+      },
+    };
+    const expectedBundleToIngest = {
+      id: '1c775f39-6cea-4b14-92f8-7843d2443af7',
+      spec_version: '2.1',
+      type: 'bundle',
+      objects: [
+        {
+          id: INCIDENT_ID,
+          spec_version: '2.1',
+          type: 'incident',
+          extensions: {
+            'extension-definition--ea279b3e-5c71-4632-ac08-831c66a786ba': {
+              extension_type: 'new-sdo',
+              id: '25cab01c-46be-48ed-832f-857d35347f15',
+              type: 'Incident',
+              is_inferred: false,
+              creator_ids: [
+                '88ec0c6a-13ce-5e39-b486-354fe4a7084f',
+              ],
+            },
+          },
+          created: '2025-02-25T08:13:45.851Z',
+          modified: '2025-05-09T10:03:11.288Z',
+          revoked: false,
+          confidence: 100,
+          lang: 'en',
+          name: 'Test Incident',
+          description: '',
+        },
+        {
+          id: 'case-incident--b52ce838-2972-51ea-a538-005b36189e19',
+          spec_version: '2.1',
+          type: 'case-incident',
+          extensions: {
+            'extension-definition--ea279b3e-5c71-4632-ac08-831c66a786ba': {
+              extension_type: 'new-sdo',
+              id: 'fe30c5e6-d68d-4cbe-a318-a046abaaacc8',
+              type: 'Case-Incident',
+            },
+          },
+          created: '2025-02-25T08:13:45.863Z',
+          name: 'Test Incident',
+          object_refs: [
+            INCIDENT_ID,
+          ],
+        },
+      ],
+    } as unknown as StixBundle;
+
+    const result = await PLAYBOOK_CONTAINER_WRAPPER_COMPONENT.executor({
+      dataInstanceId,
+      eventId: '',
+      executionId: '',
+      playbookId: '',
+      previousPlaybookNodeId: '',
+      previousStepBundle: container_wrapper_component_bundle,
+      playbookNode,
+      bundle: container_wrapper_component_bundle,
+    });
+    expect(result.bundle.objects.length).toEqual(2);
+    expect(result.bundle.objects[1].id).toEqual(expectedBundleToIngest.objects[1].id);
+    expect(result.bundle.objects[1].extensions[STIX_EXT_OCTI].type).toEqual(expectedBundleToIngest.objects[1].extensions[STIX_EXT_OCTI].type);
   });
 });
